rename RunManager to JobRunner
- move job manager to use the factory pattern
